move functions, fix tests
