+++ kitsune-http-signatures/src/header.rs
@@ -1,7 +1,12 @@
 use crate::{util::UnixTimestampExt, Error, Result, SignatureComponent};
 use base64::{engine::general_purpose, Engine};
+use chrono::{DateTime, Utc};
 use derive_builder::Builder;
-use std::time::{SystemTime, SystemTimeError};
+use http::{header::DATE, request::Parts};
+use std::{
+    cmp::min,
+    time::{Duration, SystemTime, SystemTimeError},
+};
 
 #[derive(Builder)]
 pub struct SignatureHeader<'a> {
@@ -54,6 +59,58 @@ impl<'a> SignatureHeader<'a> {
         self.signature_components
             .iter()
             .any(|component| component.as_str().to_lowercase() == "date")
+    }
+
+    /// Check if the signature is expired by checking the `(expires)` pseudo-header
+    ///
+    /// If the header isn't present, just default to considering the signature valid
+    pub fn is_expired(&self) -> bool {
+        if let Some(expires) = self.expires {
+            expires < SystemTime::now()
+        } else {
+            false
+        }
+    }
+
+    /// Check if the signature is expired really strictly
+    ///
+    /// Try to read one of the signed "created at" timestamps
+    /// If the signature doesn't contain any signed timestamps, reject the signature as invalid.
+    ///
+    /// The signature is considered expired if
+    ///
+    /// - When the `(expires)` pseudo-header tells us it is expired
+    /// - Our implementation considers the signature expired according to the setting
+    ///
+    /// Whichever is true first
+    pub fn is_expired_strict(&self, parts: &Parts, enforced_duration: Duration) -> Result<bool> {
+        let signed_created = if self
+            .signature_components
+            .contains(&SignatureComponent::Created)
+        {
+            self.created.ok_or(Error::InvalidSignatureHeader)?.into()
+        } else if self.components_include_date_header() {
+            let date_header = parts
+                .headers
+                .get(DATE)
+                .ok_or(Error::InvalidSignatureHeader)?;
+
+            DateTime::parse_from_rfc2822(date_header.to_str()?)?.with_timezone(&Utc)
+        } else {
+            return Err(Error::InvalidSignatureHeader);
+        };
+
+        // Determine when the signature is considered expired
+        let expires_at: DateTime<Utc> = {
+            let enforced_expires_at =
+                signed_created + chrono::Duration::from_std(enforced_duration)?;
+
+            self.expires.map_or(enforced_expires_at, |expires| {
+                min(expires.into(), enforced_expires_at)
+            })
+        };
+
+        Ok(Utc::now() >= expires_at)
     }
 }
